+++ firebase-ai/src/main/kotlin/com/google/firebase/ai/java/LiveSessionFutures.kt
@@ -185,17 +185,13 @@ public abstract class LiveSessionFutures internal constructor() {
   ): ListenableFuture<Unit>
 
   /**
-   * Sends audio data to the server in realtime. Check
-   * https://ai.google.dev/api/live#bidigeneratecontentrealtimeinput for details about the realtime
-   * input usage.
+   * Sends an audio input stream to the model, using the realtime API.
    * @param audio The audio data to send.
    */
   public abstract fun sendAudioRealtime(audio: InlineData): ListenableFuture<Unit>
 
   /**
-   * Sends video data to the server in realtime. Check
-   * https://ai.google.dev/api/live#bidigeneratecontentrealtimeinput for details about the realtime
-   * input usage.
+   * Sends a video input stream to the model, using the realtime API.
    * @param video The video data to send. Video MIME type could be either video or image.
    */
   public abstract fun sendVideoRealtime(video: InlineData): ListenableFuture<Unit>