+++ firebase-dataconnect/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Unreleased
 
 - [changed] Internal refactor for managing Auth and App Check tokens
-  ([#7184](https://github.com/firebase/firebase-android-sdk/pull/7184))
+  [#7485](https://github.com/firebase/firebase-android-sdk/pull/7485))
+
+# 17.1.1
+
+- [changed] Internal refactor for managing Auth and App Check tokens
+  ([#7484](https://github.com/firebase/firebase-android-sdk/pull/7484)
 
 # 17.1.0